fix(frontend): Refactor gadget property management in Canvas and App.
Moved gadget property update logic from `Canvas` to `App` for better separation of concerns. Added callbacks for selection and property updates to streamline communication between components. Improved maintainability and centralized state management for gadget interactions.
