+++ app/frontend/src/App.tsx
@@ -1,20 +1,33 @@
 import React, {useEffect, useState} from "react";
 import {offBackendEvent, onBackendEvent, ToPoint} from "./utils/wailsBridge";
 
-import {AddGadget, GetCurrentDiagramName, GetDrawData} from "../wailsjs/go/umlproject/UMLProject";
+import {
+    AddGadget,
+    GetCurrentDiagramName,
+    GetDrawData,
+    SetColorGadget,
+    SetPointGadget,
+    SetSetLayerGadget,
+    SetAttrContentGadget,
+    SetAttrSizeGadget,
+    SetAttrStyleGadget
+} from "../wailsjs/go/umlproject/UMLProject";
 import { mockAssociation, mockSelfAssociation, mockHorizontalAssociation, mockVerticalAssociation, mockSelfAssociationLeft, mockSelfAssociationUp} from "./assets/mock/ass";
 
-import {CanvasProps} from "./utils/Props";
+import {CanvasProps, GadgetProps} from "./utils/Props";
 import DrawingCanvas from "./components/Canvas";
 // import mockData from './assets/mock/gadget';
 import {GadgetPopup} from "./components/CreateGadgetPopup";
 import Toolbar from "./components/Toolbar";
 import { createAss } from "./utils/createAssociation";
+import GadgetPropertiesPanel from "./components/GadgetPropertiesPanel";
 
 const App: React.FC = () => {
     const [diagramName, setDiagramName] = useState<string | null>(null);
     const [backendData, setBackendData] = useState<CanvasProps | null>(null);
     const [showPopup, setShowPopup] = useState(false);
+    const [selectedGadget, setSelectedGadget] = useState<GadgetProps | null>(null);
+    const [selectedGadgetCount, setSelectedGadgetCount] = useState<number>(0);
 
     const handleGetDiagramName = async () => {
         try {
@@ -57,6 +70,110 @@ const App: React.FC = () => {
             setBackendData(canvasData);
         } catch (error) {
             console.error("Error loading canvas data:", error);
+        }
+    };
+
+    const handleSelectionChange = (gadget: GadgetProps | null, count: number) => {
+        setSelectedGadget(gadget);
+        setSelectedGadgetCount(count);
+    };
+
+    const handleUpdateGadgetProperty = (property: string, value: any) => {
+        if (!selectedGadget || !backendData || !backendData.gadgets) return;
+
+        console.log(`Updating property ${property} to ${value}`);
+
+        // Handle nested properties like attributes[0][0].content
+        if (property.includes('.')) {
+            const [parentProp, childProp] = property.split('.');
+            if (parentProp.startsWith('attributes')) {
+                // Parse indices from string like 'attributes[0][0]'
+                const matches = parentProp.match(/attributes(\d+):(\d+)/);
+                if (matches && matches.length === 3) {
+                    const i = parseInt(matches[1]);
+                    const j = parseInt(matches[2]);
+
+                    // console.log(i, j, childProp);
+                    if (childProp === 'content') {
+                        SetAttrContentGadget(i, j, value).then(
+                            () => {
+                                console.log("Gadget content changed");
+                                loadCanvasData();
+                            }
+                        ).catch((error) => {
+                                console.error("Error changing gadget content:", error);
+                            }
+                        );
+                    }
+                    if(childProp === 'fontSize') {
+                        SetAttrSizeGadget(i, j, value).then(
+                            () => {
+                                console.log("Gadget fontSize changed");
+                                loadCanvasData();
+                            }
+                        ).catch((error) => {
+                                console.error("Error changing gadget fontSize:", error);
+                            }
+                        );
+                    }
+                    if(childProp === 'fontStyle') {
+                        SetAttrStyleGadget(i, j, value).then(
+                            () => {
+                                console.log("Gadget fontStyle changed");
+                                loadCanvasData();
+                            }
+                        ).catch((error) => {
+                                console.error("Error changing gadget fontStyle:", error);
+                            }
+                        );
+                    }
+                }
+            }
+        } else {
+            if (property === "x") {
+                SetPointGadget(ToPoint(value, selectedGadget.y)).then(
+                    () => {
+                        console.log("Gadget moved");
+                        loadCanvasData();
+                    }
+                ).catch((error) => {
+                        console.error("Error moving gadget:", error);
+                    }
+                );
+            }
+            if (property === "y") {
+                SetPointGadget(ToPoint(selectedGadget.x, value)).then(
+                    () => {
+                        console.log("Gadget moved");
+                        loadCanvasData();
+                    }
+                ).catch((error) => {
+                        console.error("Error moving gadget:", error);
+                    }
+                );
+            }
+            if (property === "layer") {
+                SetSetLayerGadget(value).then(
+                    () => {
+                        console.log("layer changed");
+                        loadCanvasData();
+                    }
+                ).catch((error) => {
+                        console.error("Error moving gadget:", error);
+                    }
+                );
+            }
+            if (property === "color") {
+                SetColorGadget(value).then(
+                    () => {
+                        console.log("color changed");
+                        loadCanvasData();
+                    }
+                ).catch((error) => {
+                        console.error("Error moving gadget:", error);
+                    }
+                );
+            }
         }
     };
 
@@ -122,7 +239,18 @@ const App: React.FC = () => {
                     onClose={() => setShowPopup(false)}
                 />
             )}
-            <DrawingCanvas backendData={backendData} reloadBackendData={loadCanvasData} />
+            <DrawingCanvas
+                backendData={backendData}
+                reloadBackendData={loadCanvasData}
+                onSelectionChange={handleSelectionChange}
+                onUpdateGadgetProperty={handleUpdateGadgetProperty}
+            />
+            {selectedGadgetCount === 1 && (
+                <GadgetPropertiesPanel
+                    selectedGadget={selectedGadget}
+                    updateGadgetProperty={handleUpdateGadgetProperty}
+                />
+            )}
         </div>
     );
 };
